+++ src/test/java/org/springframework/data/redis/connection/DefaultStringRedisConnectionTests.java
@@ -1263,15 +1263,15 @@ public class DefaultStringRedisConnectionTests {
 
 	@Test
 	public void testZInterStoreAggWeightsBytes() {
-		doReturn(5l).when(nativeConnection).zInterStore(fooBytes, Aggregate.MAX, new int[0], fooBytes);
-		actual.add(connection.zInterStore(fooBytes, Aggregate.MAX, new int[0], fooBytes));
+		doReturn(5l).when(nativeConnection).zInterStore(fooBytes, Aggregate.MAX, new double[0], fooBytes);
+		actual.add(connection.zInterStore(fooBytes, Aggregate.MAX, new double[0], fooBytes));
 		verifyResults(Arrays.asList(new Object[] { 5l }));
 	}
 
 	@Test
 	public void testZInterStoreAggWeights() {
-		doReturn(5l).when(nativeConnection).zInterStore(fooBytes, Aggregate.MAX, new int[0], fooBytes);
-		actual.add(connection.zInterStore(foo, Aggregate.MAX, new int[0], foo));
+		doReturn(5l).when(nativeConnection).zInterStore(fooBytes, Aggregate.MAX, new double[0], fooBytes);
+		actual.add(connection.zInterStore(foo, Aggregate.MAX, new double[0], foo));
 		verifyResults(Arrays.asList(new Object[] { 5l }));
 	}
 
@@ -1543,15 +1543,15 @@ public class DefaultStringRedisConnectionTests {
 
 	@Test
 	public void testZUnionStoreAggWeightsBytes() {
-		doReturn(5l).when(nativeConnection).zUnionStore(fooBytes, Aggregate.MAX, new int[0], fooBytes);
-		actual.add(connection.zUnionStore(fooBytes, Aggregate.MAX, new int[0], fooBytes));
+		doReturn(5l).when(nativeConnection).zUnionStore(fooBytes, Aggregate.MAX, new double[0], fooBytes);
+		actual.add(connection.zUnionStore(fooBytes, Aggregate.MAX, new double[0], fooBytes));
 		verifyResults(Arrays.asList(new Object[] { 5l }));
 	}
 
 	@Test
 	public void testZUnionStoreAggWeights() {
-		doReturn(5l).when(nativeConnection).zUnionStore(fooBytes, Aggregate.MAX, new int[0], fooBytes);
-		actual.add(connection.zUnionStore(foo, Aggregate.MAX, new int[0], foo));
+		doReturn(5l).when(nativeConnection).zUnionStore(fooBytes, Aggregate.MAX, new double[0], fooBytes);
+		actual.add(connection.zUnionStore(foo, Aggregate.MAX, new double[0], foo));
 		verifyResults(Arrays.asList(new Object[] { 5l }));
 	}
